Moving LeafHasher logic to a separate file. Minor code reordering in OnRamp files (constructor after struct)
