+++ core/services/ocr2/plugins/ccip/internal/ccipdata/v1_0_0/onramp.go
@@ -2,17 +2,14 @@ package v1_0_0
 
 import (
 	"context"
+	"errors"
 	"fmt"
-	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/math"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/pkg/errors"
 
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/client"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
-	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/utils"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp_1_0_0"
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/abihelpers"
@@ -25,75 +22,7 @@ import (
 
 const (
 	CCIPSendRequestedEventName = "CCIPSendRequested"
-	MetaDataHashPrefix         = "EVM2EVMMessageEvent"
 )
-
-var LeafDomainSeparator = [1]byte{0x00}
-
-type LeafHasher struct {
-	metaDataHash [32]byte
-	ctx          hashlib.Ctx[[32]byte]
-	onRamp       *evm_2_evm_onramp_1_0_0.EVM2EVMOnRamp
-}
-
-func GetMetaDataHash[H hashlib.Hash](ctx hashlib.Ctx[H], prefix [32]byte, sourceChainSelector uint64, onRampId common.Address, destChainSelector uint64) H {
-	paddedOnRamp := common.BytesToHash(onRampId[:])
-	return ctx.Hash(utils.ConcatBytes(prefix[:], math.U256Bytes(big.NewInt(0).SetUint64(sourceChainSelector)), math.U256Bytes(big.NewInt(0).SetUint64(destChainSelector)), paddedOnRamp[:]))
-}
-
-func NewLeafHasher(sourceChainSelector uint64, destChainSelector uint64, onRampId common.Address, ctx hashlib.Ctx[[32]byte], onRamp *evm_2_evm_onramp_1_0_0.EVM2EVMOnRamp) *LeafHasher {
-	return &LeafHasher{
-		metaDataHash: GetMetaDataHash(ctx, ctx.Hash([]byte(MetaDataHashPrefix)), sourceChainSelector, onRampId, destChainSelector),
-		ctx:          ctx,
-		onRamp:       onRamp,
-	}
-}
-
-func (t *LeafHasher) HashLeaf(log types.Log) ([32]byte, error) {
-	message, err := t.onRamp.ParseCCIPSendRequested(log)
-	if err != nil {
-		return [32]byte{}, err
-	}
-	encodedTokens, err := utils.ABIEncode(
-		`[
-{"components": [{"name":"token","type":"address"},{"name":"amount","type":"uint256"}], "type":"tuple[]"}]`, message.Message.TokenAmounts)
-	if err != nil {
-		return [32]byte{}, err
-	}
-
-	packedValues, err := utils.ABIEncode(
-		`[
-{"name": "leafDomainSeparator","type":"bytes1"},
-{"name": "metadataHash", "type":"bytes32"},
-{"name": "sequenceNumber", "type":"uint64"},
-{"name": "nonce", "type":"uint64"},
-{"name": "sender", "type":"address"},
-{"name": "receiver", "type":"address"},
-{"name": "dataHash", "type":"bytes32"},
-{"name": "tokenAmountsHash", "type":"bytes32"},
-{"name": "gasLimit", "type":"uint256"},
-{"name": "strict", "type":"bool"},
-{"name": "feeToken","type": "address"},
-{"name": "feeTokenAmount","type": "uint256"}
-]`,
-		LeafDomainSeparator,
-		t.metaDataHash,
-		message.Message.SequenceNumber,
-		message.Message.Nonce,
-		message.Message.Sender,
-		message.Message.Receiver,
-		t.ctx.Hash(message.Message.Data),
-		t.ctx.Hash(encodedTokens),
-		message.Message.GasLimit,
-		message.Message.Strict,
-		message.Message.FeeToken,
-		message.Message.FeeTokenAmount,
-	)
-	if err != nil {
-		return [32]byte{}, err
-	}
-	return t.ctx.Hash(packedValues), nil
-}
 
 var _ ccipdata.OnRampReader = &OnRamp{}
 
@@ -107,44 +36,6 @@ type OnRamp struct {
 	sendRequestedEventSig      common.Hash
 	sendRequestedSeqNumberWord int
 	filters                    []logpoller.Filter
-}
-
-func (o *OnRamp) Address() (cciptypes.Address, error) {
-	return cciptypes.Address(o.onRamp.Address().String()), nil
-}
-
-func (o *OnRamp) GetDynamicConfig() (cciptypes.OnRampDynamicConfig, error) {
-	if o.onRamp == nil {
-		return cciptypes.OnRampDynamicConfig{}, fmt.Errorf("onramp not initialized")
-	}
-	legacyDynamicConfig, err := o.onRamp.GetDynamicConfig(nil)
-	if err != nil {
-		return cciptypes.OnRampDynamicConfig{}, err
-	}
-	return cciptypes.OnRampDynamicConfig{
-		Router:                            cciptypes.Address(legacyDynamicConfig.Router.String()),
-		MaxNumberOfTokensPerMsg:           legacyDynamicConfig.MaxTokensLength,
-		DestGasOverhead:                   0,
-		DestGasPerPayloadByte:             0,
-		DestDataAvailabilityOverheadGas:   0,
-		DestGasPerDataAvailabilityByte:    0,
-		DestDataAvailabilityMultiplierBps: 0,
-		PriceRegistry:                     cciptypes.Address(legacyDynamicConfig.PriceRegistry.String()),
-		MaxDataBytes:                      legacyDynamicConfig.MaxDataSize,
-		MaxPerMsgGasLimit:                 uint32(legacyDynamicConfig.MaxGasLimit),
-	}, nil
-}
-
-func (o *OnRamp) GetLastUSDCMessagePriorToLogIndexInTx(ctx context.Context, logIndex int64, txHash common.Hash) ([]byte, error) {
-	return nil, errors.New("USDC not supported in < 1.2.0")
-}
-
-func (o *OnRamp) Close(qopts ...pg.QOpt) error {
-	return logpollerutil.UnregisterLpFilters(o.lp, o.filters, qopts...)
-}
-
-func (o *OnRamp) RegisterFilters(qopts ...pg.QOpt) error {
-	return logpollerutil.RegisterLpFilters(o.lp, o.filters, qopts...)
 }
 
 func NewOnRamp(lggr logger.Logger, sourceSelector, destSelector uint64, onRampAddress common.Address, sourceLP logpoller.LogPoller, source client.Client) (*OnRamp, error) {
@@ -175,37 +66,29 @@ func NewOnRamp(lggr logger.Logger, sourceSelector, destSelector uint64, onRampAd
 	}, nil
 }
 
-func (o *OnRamp) logToMessage(log types.Log) (*cciptypes.EVM2EVMMessage, error) {
-	msg, err := o.onRamp.ParseCCIPSendRequested(log)
+func (o *OnRamp) Address() (cciptypes.Address, error) {
+	return cciptypes.Address(o.onRamp.Address().String()), nil
+}
+
+func (o *OnRamp) GetDynamicConfig() (cciptypes.OnRampDynamicConfig, error) {
+	if o.onRamp == nil {
+		return cciptypes.OnRampDynamicConfig{}, fmt.Errorf("onramp not initialized")
+	}
+	legacyDynamicConfig, err := o.onRamp.GetDynamicConfig(nil)
 	if err != nil {
-		return nil, err
+		return cciptypes.OnRampDynamicConfig{}, err
 	}
-	h, err := o.leafHasher.HashLeaf(log)
-	if err != nil {
-		return nil, err
-	}
-	tokensAndAmounts := make([]cciptypes.TokenAmount, len(msg.Message.TokenAmounts))
-	for i, tokenAndAmount := range msg.Message.TokenAmounts {
-		tokensAndAmounts[i] = cciptypes.TokenAmount{
-			Token:  cciptypes.Address(tokenAndAmount.Token.String()),
-			Amount: tokenAndAmount.Amount,
-		}
-	}
-	return &cciptypes.EVM2EVMMessage{
-		SequenceNumber:      msg.Message.SequenceNumber,
-		GasLimit:            msg.Message.GasLimit,
-		Nonce:               msg.Message.Nonce,
-		MessageID:           msg.Message.MessageId,
-		SourceChainSelector: msg.Message.SourceChainSelector,
-		Sender:              cciptypes.Address(msg.Message.Sender.String()),
-		Receiver:            cciptypes.Address(msg.Message.Receiver.String()),
-		Strict:              msg.Message.Strict,
-		FeeToken:            cciptypes.Address(msg.Message.FeeToken.String()),
-		FeeTokenAmount:      msg.Message.FeeTokenAmount,
-		Data:                msg.Message.Data,
-		TokenAmounts:        tokensAndAmounts,
-		SourceTokenData:     make([][]byte, len(msg.Message.TokenAmounts)), // Always empty in 1.0
-		Hash:                h,
+	return cciptypes.OnRampDynamicConfig{
+		Router:                            cciptypes.Address(legacyDynamicConfig.Router.String()),
+		MaxNumberOfTokensPerMsg:           legacyDynamicConfig.MaxTokensLength,
+		DestGasOverhead:                   0,
+		DestGasPerPayloadByte:             0,
+		DestDataAvailabilityOverheadGas:   0,
+		DestGasPerDataAvailabilityByte:    0,
+		DestDataAvailabilityMultiplierBps: 0,
+		PriceRegistry:                     cciptypes.Address(legacyDynamicConfig.PriceRegistry.String()),
+		MaxDataBytes:                      legacyDynamicConfig.MaxDataSize,
+		MaxPerMsgGasLimit:                 uint32(legacyDynamicConfig.MaxGasLimit),
 	}, nil
 }
 
@@ -243,4 +126,50 @@ func (o *OnRamp) RouterAddress() (cciptypes.Address, error) {
 		return "", err
 	}
 	return cciptypes.Address(config.Router.String()), nil
+}
+
+func (o *OnRamp) GetLastUSDCMessagePriorToLogIndexInTx(ctx context.Context, logIndex int64, txHash common.Hash) ([]byte, error) {
+	return nil, errors.New("USDC not supported in < 1.2.0")
+}
+
+func (o *OnRamp) Close(qopts ...pg.QOpt) error {
+	return logpollerutil.UnregisterLpFilters(o.lp, o.filters, qopts...)
+}
+
+func (o *OnRamp) RegisterFilters(qopts ...pg.QOpt) error {
+	return logpollerutil.RegisterLpFilters(o.lp, o.filters, qopts...)
+}
+
+func (o *OnRamp) logToMessage(log types.Log) (*cciptypes.EVM2EVMMessage, error) {
+	msg, err := o.onRamp.ParseCCIPSendRequested(log)
+	if err != nil {
+		return nil, err
+	}
+	h, err := o.leafHasher.HashLeaf(log)
+	if err != nil {
+		return nil, err
+	}
+	tokensAndAmounts := make([]cciptypes.TokenAmount, len(msg.Message.TokenAmounts))
+	for i, tokenAndAmount := range msg.Message.TokenAmounts {
+		tokensAndAmounts[i] = cciptypes.TokenAmount{
+			Token:  cciptypes.Address(tokenAndAmount.Token.String()),
+			Amount: tokenAndAmount.Amount,
+		}
+	}
+	return &cciptypes.EVM2EVMMessage{
+		SequenceNumber:      msg.Message.SequenceNumber,
+		GasLimit:            msg.Message.GasLimit,
+		Nonce:               msg.Message.Nonce,
+		MessageID:           msg.Message.MessageId,
+		SourceChainSelector: msg.Message.SourceChainSelector,
+		Sender:              cciptypes.Address(msg.Message.Sender.String()),
+		Receiver:            cciptypes.Address(msg.Message.Receiver.String()),
+		Strict:              msg.Message.Strict,
+		FeeToken:            cciptypes.Address(msg.Message.FeeToken.String()),
+		FeeTokenAmount:      msg.Message.FeeTokenAmount,
+		Data:                msg.Message.Data,
+		TokenAmounts:        tokensAndAmounts,
+		SourceTokenData:     make([][]byte, len(msg.Message.TokenAmounts)), // Always empty in 1.0
+		Hash:                h,
+	}, nil
 }
